Remove unused EXCALIDRAW_ASSET_PATH
We are bundling all necessary assets into JS/CSS so that we no
longer need to statically serve an `excalidraw-dist/`. In case
of a package upgrade in the future breaking this contract, we
could either point the URL to a non-existent path to catch errors
or let it fall back to the default unpkg mirror. We choose the
latter.